Implementation of an error correcting code
This is an error correcting hamming code that uses an (8,4) parity
matrix.

To test, uncomment the random bit flipping that simulates errors and
run it with “declaration.txt”. It will output the encoded and decoded
files. By diff-ing the decoded and original, you can see that it will
correct any single error that occurs.